Reference

Input data types

SmartAMLSim.TransactionDistribution — Type

TransactionDistribution

Type to hold the information about distribution of transactions.

SmartAMLSim.Entity — Type

Entity

Type to hold the information about distribution of transactions.

SmartAMLSim.AMLInternalRule — Type

AMLInternalRule

Type representing internal AML rule used by a financial institution.

side::Char - either D or C, for Debit and Credit correspondingly; Credit rule monitor transactions incoming to the internal customer

business_type::String - a keyword matching business type of entities, e.g. ATM, International

amount::Number - threshold value for triggering an alert with a given rule

SmartAMLSim.AMLExternalRule — Type

AMLExternalRule

Type representing external AML systems maintained by external financial institutions.

business_type::String - a keyword matching business type of entities, e.g. ATM, International

detection_multiplier::Float64 - value of detection multiplier which determines detection chance linearly increasing with the value of outgoing transactions of particular business_type

Data loading functions

SmartAMLSim.load_transactions_distributions — Function

load_transactions_distributions(filename::String, sep::Char=' ')

Loads distributions of transactions from provided filename into Vector of TransactionDistribution objects. sep is used to split every line of the file.

SmartAMLSim.load_entities — Function

load_entities(filename::String, sep::Char=',')

Loads simulation entities from provided filename into dictionary with entity's ID as a key and Entity object as value. sep is used to split every line of the input file.

SmartAMLSim.load_internal_aml_specification — Function

load_internal_aml_specification(filename::String, sep::Char=',')

Loads internal AML rules from provided filename into Vector of AMLInternalRule objects. sep is used to split every line of the input file.

SmartAMLSim.load_external_aml_specification — Function

load_external_aml_specification(filename::String, sep::Char=',')

Loads external AML rules from provided filename into Vector of AMLExternalRule objects. sep is used to split every line of the input file.

Simulation environment

SmartAMLSim.SmartAMLSimEnv — Type

SmartAMLSimEnv

Simulation environment holding information about distributions of customers' transactions.

SmartAMLSim.initialize_environment — Function

initialize_environment(
    transactions_filename::String,
    entities_filename::String,
    transactions_sep::Char=' ',
    other_sep::Char=',',
    verbose::Bool=false
)

Parses input data and initiliaze simulation's environment object for further processing.

SmartAMLSim.validate_env — Function

validate_env(env::SmartAMLSimEnv, return_internal_ids::Bool=false, verbose::Bool=false)

Validate integrity between enitites list and transactions; optionally, returns list of internal customers IDs for initialization of FinancialInstitutionAgent

SmartAMLSim.generate_transactions — Function

generate_transactions(env::SmartAMLSimEnv)

Generates customers' transactions for an epoch. Transactions are clipped at 0 as the value can't be negative.

Financial institution agent

SmartAMLSim.FinancialInstitutionAgent — Type

FinancialInstitutionAgent

Representation of financial institution monitoring transactional network and detecting money laundering with internal AML system and fixed investigation capacity.

SmartAMLSim.initialize_fi_agent — Function

initialize_fi_agent(env::SmartAMLSimEnv,
    unique_internals_from_trxs::Vector{Int},
    aml_internal_spec_filename::String,
    capacity::Int,
    file_sep::Char=',',
    verbose::Bool=false)

Function initializes FinancialInstitutionAgentloads by loading internal AML rules from file and mapping transactions to AML rules for improved processing performance.

SmartAMLSim.calculate_alerts — Function

calculate_alerts(fi::FinancialInstitutionAgent,
    epoch_trxs_vec::Vector{Float64},
    calculate_probabilities::Bool=true)

Counts how many customers breached 1,2,...,n internal AML rules. By defualt, FI's capacity is used to calculate probability of detection for each number of rules breached.

Launderer agent

SmartAMLSim.LaunderingAgent — Type

LaunderingAgent

Representation of money laundering customer of financial institution. Based on the attitude to risk launderer expose a particular utility function. Launderer also has a particular specification of external AML systems.

SmartAMLSim.initialize_laundering_agent — Function

initialize_laundering_agent(launderer_risk_attitude::String,
aml_external_spec_filename::String,
utility_funcs_dict::Dict{String,Function}),
sep=',',
verbose=false)

Validates the risk attitude string and translates it to launderer's utility function. External AML rules are loaded from file and validated. Returns LaunderingAgent object.

SmartAMLSim.optimize_launderer_decision — Function

optimize_launderer_decision(agent::LaunderingAgent,
    fi::FinancialInstitutionAgent,
    prob_map::Dict{Int,Float64},
    verbose::Bool=false)

Runs multiple optimisation models for combinations of breached internal AML rules and input probability of detection for each number of rules breached. Finds transactions specification which maximize expected utility of laundering agent in given epoch.

SmartAMLSim.optimize_single_problem — Function

optimize_single_problem(agent::LaunderingAgent,
    fi::FinancialInstitutionAgent,
    rules_breached::Vector{Int},
    p_internal::Float64)

Build and optimize single scenario (combination of internal rules breached) for a given epoch.

Simulation run

SmartAMLSim.run_epoch — Function

run_epoch(env::SmartAMLSimEnv, launderer::LaunderingAgent, fi::FinancialInstitutionAgent)

Runs one epoch of the simulation which constitutes of:

generating customers' transactions based on distributions

calculating alerts and detection probabilities by financial institution

optimiziation of launderer decision

SmartAMLSim.run_n_epochs — Function

run_n_epochs(n::Int, env::SmartAMLSimEnv, launderer::LaunderingAgent, fi::FinancialInstitutionAgent)

Runs n consecutive epochs and returns statistics for analysis.